+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Improved
 - Improve alt text judgment in keyword checklist
+- Check for keyword in first paragraph of body, main, and article tags
 
 ## 3.5.4 - 2019-01-28